fix(app): rename incoming call to call invite
